fix: 드래그 시 모든 셀 리렌더링 되는 것 개선
